Stats Test Updated

+++ stats-test.cpp
@@ -17,7 +17,7 @@ TEST_CASE("reports average, minimum and maximum") {
 TEST_CASE("reports average, minimum and maximum for array containing atleast one NaN input") 
 {
     auto computedStats = Statistics::ComputeStatistics({2.1, 7.9,2.5,NAN, 4.5,NAN});
-    double epsilon = 0.001;
+    float epsilon = 0.001;
     REQUIRE(std::abs(computedStats.average - 4.25) < epsilon);
     REQUIRE(std::abs(computedStats.max - 7.9) < epsilon);
     REQUIRE(std::abs(computedStats.min - 2.1) < epsilon);
